✏️ Change type of price in dto

+++ src/DTO/box/CreateBoxRequest.php
@@ -8,13 +8,13 @@ use Symfony\Contracts\Translation\TranslatorInterface;
 class CreateBoxRequest
 {
     public string $name;
-    public int $price;
+    public float $price;
     public int $type;
     public int $quantity;
     public int $brawler_quantity;
     public array $brawlers_in_box;
 
-    public function __construct(string $name, int $price, int $type, int $quantity, int $brawler_quantity, array $brawlers_in_box)
+    public function __construct(string $name, float $price, int $type, int $quantity, int $brawler_quantity, array $brawlers_in_box)
     {
         $this->name = $name;
         $this->price = $price;
